exercises/2-intermediate/config/workflow-controller-python/requirements.txt
kopf>=1.35.4
kubernetes>=26.1.0
prometheus-client>=0.17.0
boto3>=1.26.137
aiohttp>=3.8.4
pyyaml>=6.0